@@ -2,49 +2,48 @@
   "name": "go-barber",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:3000",
   "dependencies": {
     "@rocketseat/unform": "^1.6.2",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "axios": "^0.19.2",
-    "babel-plugin-root-import": "^6.4.1",
-    "customize-cra": "^0.9.1",
-    "date-fns": "^2.0.0-beta.5",
-    "date-fns-tz": "^1.0.10",
-    "history": "^4.10.1",
-    "immer": "^6.0.1",
-    "polished": "^3.4.4",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.0",
-    "react-app-rewired": "^2.1.5",
-    "react-dom": "^16.13.0",
-    "react-icons": "^3.9.0",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.25.0",
+    "customize-cra": "^1.0.0",
+    "date-fns": "^2.28.0",
+    "date-fns-tz": "^1.2.2",
+    "history": "^5.2.0",
+    "polished": "^4.1.4",
+    "prop-types": "^15.8.1",
+    "react": "^17.0.2",
+    "react-app-rewired": "^2.1.11",
+    "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
     "react-perfect-scrollbar": "^1.5.8",
-    "react-redux": "^7.2.0",
-    "react-redux-toastr": "^7.6.4",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.0",
+    "react-redux": "^7.2.6",
+    "react-redux-toastr": "^7.6.8",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
     "reactotron-react-js": "^3.3.7",
-    "reactotron-redux": "^3.1.2",
+    "reactotron-redux": "^3.1.3",
     "reactotron-redux-saga": "^4.2.3",
-    "redux": "^4.0.5",
+    "redux": "^4.1.2",
     "redux-persist": "^6.0.0",
     "redux-saga": "^1.1.3",
-    "reduxsauce": "^1.1.2",
+    "reduxsauce": "^1.2.1",
     "seamless-immutable": "^7.1.4",
-    "styled-components": "^5.0.1",
-    "yup": "^0.28.3"
+    "styled-components": "^5.3.3",
+    "yup": "^0.32.11"
   },
   "scripts": {
     "start": "react-app-rewired start",
     "build": "react-app-rewired build",
-    "test": "react-app-rewired test",
-    "eject": "react-scripts eject"
+    "test": "react-app-rewired test"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -59,33 +58,20 @@
     ]
   },
   "devDependencies": {
-    "babel-eslint": "10.0.3",
-    "babel-plugin-import": "^1.13.0",
-    "@commitlint/cli": "^8.3.5",
-    "@commitlint/config-conventional": "^8.3.4",
-    "commitizen": "^4.0.3",
-    "cz-conventional-changelog": "3.1.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-import-resolver-babel-plugin-root-import": "^1.1.1",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "husky": "^4.2.3",
-    "prettier": "^1.19.1"
-  },
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
-    }
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
-      "prepare-commit-msg": "exec < /dev/tty && git cz --hook || true"
-    }
+    "@types/prop-types": "^15.7.4",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@typescript-eslint/parser": "^5.10.2",
+    "babel-plugin-root-import": "^6.6.0",
+    "eslint": "^8.8.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "prettier": "^2.5.1",
+    "typescript": "^4.5.5"
   }
 }
