@@ -1,57 +1,49 @@
 module.exports = {
   env: {
     browser: true,
-    es6: true,
-    jest: true,
+    es2021: true,
   },
-  root: true,
-  extends: ['airbnb', 'prettier', 'prettier/react'],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly',
-  },
-  parser: 'babel-eslint',
+  extends: [
+    'plugin:react/recommended',
+    'airbnb',
+    'prettier',
+    'plugin:import/recommended',
+    'plugin:import/typescript',
+  ],
+  parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaFeatures: {
       jsx: true,
     },
-    ecmaVersion: 2018,
+    ecmaVersion: 'latest',
     sourceType: 'module',
   },
-  plugins: ['react', 'prettier', 'jsx-a11y', 'react-hooks'],
+  plugins: ['react', '@typescript-eslint', 'prettier', 'import'],
   rules: {
+    'import/no-unresolved': 'error',
     'prettier/prettier': 'error',
     'react/jsx-filename-extension': ['warn', { extensions: ['.js', '.jsx'] }],
     'import/prefer-default-export': 'off',
     'no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
-    'react/jsx-one-expression-per-line': 'off',
-    'global-require': 'off',
-    'react-native/no-raw-text': 'off',
-    'no-param-reassign': 'off',
-    'no-underscore-dangle': 'off',
-    camelcase: 'off',
     'no-console': ['error', { allow: ['tron'] }],
-    'react-hooks/rules-of-hooks': 'error',
-    'react-hooks/exhaustive-deps': 'warn',
-    'react/state-in-constructor': 'off',
-    'import/no-cycle': 'off',
-    'react/jsx-fragments': 'off',
-    'jsx-a11y/label-has-associated-control': 'off',
     'react/jsx-props-no-spreading': 'off',
-    'react/forbid-prop-types': 'off',
-    'jsx-a11y/click-events-have-key-events': 'off',
-    'jsx-a11y/no-static-element-interactions': 'off',
-    'prefer-template': 'off',
-    eqeqeq: 'off',
-    'react/no-array-index-key': 'off',
-    'no-console': 'off',
-    'no-irregular-whitespace': 'off',
-    'no-alert': 'off',
+    camelcase: 'off',
+    'react/react-in-jsx-scope': 'off',
+    'react/function-component-definition': [
+      2,
+      {
+        namedComponents: 'arrow-function',
+        unnamedComponents: 'arrow-function',
+      },
+    ],
   },
   settings: {
+    'import/parsers': {
+      '@typescript-eslint/parser': ['.ts', '.tsx'],
+    },
     'import/resolver': {
-      'babel-plugin-root-import': {
-        rootPathSuffix: 'src',
+      typescript: {
+        alwaysTryTypes: true,
       },
     },
   },
