@@ -1,4 +1,5 @@
-import React, { useState, useEffect, useMemo } from 'react';
+/* eslint-disable import/no-duplicates */
+import React, { useState, useEffect, useMemo, useCallback } from 'react';
 import { MdNotifications } from 'react-icons/md';
 import { parseISO, formatDistance } from 'date-fns';
 import pt from 'date-fns/locale/pt';
@@ -13,49 +14,50 @@ import {
   Scroll,
 } from './styles';
 
-export default function Notifications() {
+const Notifications = () => {
   const [visible, setVisible] = useState(false);
   const [notifications, setNotifications] = useState([]);
 
   const hasUnread = useMemo(
-    () => !!notifications.find(notification => notification.read === false),
+    () => !!notifications.find((notification) => notification.read === false),
     [notifications]
   );
 
+  const loadNotifications = useCallback(async () => {
+    const response = await api.get('notifications');
+
+    const data = response.data.map((notification) => ({
+      ...notification,
+      timeDistance: formatDistance(
+        parseISO(notification.created_at),
+        new Date(),
+        {
+          addSuffix: true,
+          locale: pt,
+        }
+      ),
+    }));
+
+    setNotifications(data);
+  }, []);
+
   useEffect(() => {
-    async function loadNotifications() {
-      const response = await api.get('notifications');
-
-      const data = response.data.map(notification => ({
-        ...notification,
-        timeDistance: formatDistance(
-          parseISO(notification.created_at),
-          new Date(),
-          {
-            addSuffix: true,
-            locale: pt,
-          }
-        ),
-      }));
-
-      setNotifications(data);
-    }
     loadNotifications();
   }, []);
 
-  function handleToggleVisible() {
+  const handleToggleVisible = () => {
     setVisible(!visible);
-  }
+  };
 
-  async function handleMarkAsread(id) {
+  const handleMarkAsread = async (id) => {
     await api.patch(`notifications/${id}`);
 
     setNotifications(
-      notifications.map(notification =>
+      notifications.map((notification) =>
         notification.id === id ? { ...notification, read: true } : notification
       )
     );
-  }
+  };
 
   return (
     <Container>
@@ -64,7 +66,7 @@ export default function Notifications() {
       </Badge>
       <NotificationList visible={visible}>
         <Scroll>
-          {notifications.map(notification => (
+          {notifications.map((notification) => (
             <Notification key={notification.id} unread={!notification.read}>
               <p>Voce possui um novo agendamento para amannha</p>
               <time>ha 2 dias</time>
@@ -82,4 +84,6 @@ export default function Notifications() {
       </NotificationList>
     </Container>
   );
-}
+};
+
+export default Notifications;
