@@ -3,13 +3,13 @@ import PropTypes from 'prop-types';
 
 import { Wrapper, Content } from './styles';
 
-export default function AuthLayout({ children }) {
-  return (
-    <Wrapper>
-      <Content>{children}</Content>
-    </Wrapper>
-  );
-}
+const AuthLayout = ({ children }) => (
+  <Wrapper>
+    <Content>{children}</Content>
+  </Wrapper>
+);
+
+export default AuthLayout;
 
 AuthLayout.propTypes = {
   children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]).isRequired,
