@@ -1,17 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { Wrapper } from './styles';
 import Header from '~/components/Header';
+import { Wrapper } from './styles';
 
-export default function DefaultLayout({ children }) {
-  return (
-    <Wrapper>
-      <Header />
-      {children}
-    </Wrapper>
-  );
-}
+const DefaultLayout = ({ children }) => (
+  <Wrapper>
+    <Header />
+    {children}
+  </Wrapper>
+);
+
+export default DefaultLayout;
 
 DefaultLayout.propTypes = {
   children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]).isRequired,
