@@ -1,90 +1,97 @@
-import React, { useState, useMemo, useEffect } from 'react';
-import { MdChevronLeft, MdChevronRight } from 'react-icons/md';
-import {
-  format,
-  subDays,
-  addDays,
-  setHours,
-  setMinutes,
-  setSeconds,
-  setMilliseconds,
-  isBefore,
-  isEqual,
-  parseISO,
-} from 'date-fns';
-import { utcToZonedTime } from 'date-fns-tz';
-import pt from 'date-fns/locale/pt';
+// /* eslint-disable import/no-duplicates */
+// import React, { useState, useMemo, useEffect } from 'react';
+// import { MdChevronLeft, MdChevronRight } from 'react-icons/md';
+// import {
+//   format,
+//   subDays,
+//   addDays,
+//   setHours,
+//   setMinutes,
+//   setSeconds,
+//   setMilliseconds,
+//   isBefore,
+//   isEqual,
+//   parseISO,
+// } from 'date-fns';
+// import pt from 'date-fns/locale/pt';
 
-import { Container, Time } from './styles';
+// import { utcToZonedTime } from 'date-fns-tz';
 
-import api from '~/services/api';
+// import api from '~/services/api';
+// import { Container, Time } from './styles';
 
-const range = [8, 9, 10, 11, 13, 14, 15, 16, 17, 18];
+// const range = [8, 9, 10, 11, 13, 14, 15, 16, 17, 18];
 
-export default function Dashboard() {
-  const [schedule, setSchedule] = useState([]);
-  const [date, setDate] = useState(new Date());
+// const Dashboard = () => {
+//   const [schedule, setSchedule] = useState([]);
+//   const [date, setDate] = useState(new Date());
 
-  const dateFormatted = useMemo(
-    () => format(date, "d 'de' MMMM", { locale: pt }),
-    [date]
-  );
+//   const dateFormatted = useMemo(
+//     () => format(date, "d 'de' MMMM", { locale: pt }),
+//     [date]
+//   );
 
-  function handlePrevDate() {
-    setDate(subDays(date, 1));
-  }
+//   function handlePrevDate() {
+//     setDate(subDays(date, 1));
+//   }
 
-  function handleNextDate() {
-    setDate(addDays(date, 1));
-  }
+//   function handleNextDate() {
+//     setDate(addDays(date, 1));
+//   }
 
-  useEffect(() => {
-    async function loadSchedule() {
-      const res = await api.get('schedule', { params: { date } });
+//   useEffect(() => {
+//     async function loadSchedule() {
+//       const res = await api.get('schedule', { params: { date } });
 
-      const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-      const data = range.map(hour => {
-        const checkDate = setMilliseconds(
-          setSeconds(setMinutes(setHours(date, hour), 0), 0),
-          0
-        );
-        const compareDate = utcToZonedTime(checkDate, timezone);
+//       const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
+//       const data = range.map((hour) => {
+//         const checkDate = setMilliseconds(
+//           setSeconds(setMinutes(setHours(date, hour), 0), 0),
+//           0
+//         );
+//         const compareDate = utcToZonedTime(checkDate, timezone);
 
-        return {
-          time: `${hour}:00h`,
-          past: isBefore(checkDate, new Date()),
-          appointment: res.data.find(a =>
-            isEqual(parseISO(a.date), compareDate)
-          ),
-        };
-      });
-      setSchedule(data);
-    }
-    loadSchedule();
-  }, [date]);
+//         return {
+//           time: `${hour}:00h`,
+//           past: isBefore(checkDate, new Date()),
+//           appointment: res.data.find((a) =>
+//             isEqual(parseISO(a.date), compareDate)
+//           ),
+//         };
+//       });
+//       setSchedule(data);
+//     }
+//     loadSchedule();
+//   }, [date]);
 
-  return (
-    <Container>
-      <header>
-        <button type="button" onClick={handlePrevDate}>
-          <MdChevronLeft size={36} color="#fff" />
-        </button>
-        <strong>{dateFormatted}</strong>
-        <button type="button" onClick={handleNextDate}>
-          <MdChevronRight size={36} color="#fff" />
-        </button>
-      </header>
+//   return (
+//     <Container>
+//       <header>
+//         <button type="button" onClick={handlePrevDate}>
+//           <MdChevronLeft size={36} color="#fff" />
+//         </button>
+//         <strong>{dateFormatted}</strong>
+//         <button type="button" onClick={handleNextDate}>
+//           <MdChevronRight size={36} color="#fff" />
+//         </button>
+//       </header>
 
-      <ul>
-        {schedule.map(time => (
-          <Time key={time.time} past={time.past} available={!time.appointment}>
-            <strong>{time.time}</strong>
-            <span>
-              {time.appointment ? time.appointment.user.name : 'Em aberto'}
-            </span>
-          </Time>
-        ))}
-      </ul>
-    </Container>
-  );
-}
+//       <ul>
+//         {schedule.map((time) => (
+//           <Time key={time.time} past={time.past} available={!time.appointment}>
+//             <strong>{time.time}</strong>
+//             <span>
+//               {time.appointment ? time.appointment.user.name : 'Em aberto'}
+//             </span>
+//           </Time>
+//         ))}
+//       </ul>
+//     </Container>
+//   );
+// };
+
+// export default Dashboard;
+
+const Dashboard = () => <div />;
+
+export default Dashboard;
