@@ -35,11 +35,11 @@ export const Time = styled.li`
   padding: 20px;
   border-radius: 4px;
   background: #fff;
-  opacity: ${props => (props.past ? 0.6 : 1)};
+  opacity: ${(props) => (props.past ? 0.6 : 1)};
 
   strong {
     display: block;
-    color: ${props => (props.available ? '#999' : '#7159c1')};
+    color: ${(props) => (props.available ? '#999' : '#7159c1')};
     font-size: 20px;
     font-weight: normal;
   }
@@ -47,6 +47,6 @@ export const Time = styled.li`
   span {
     display: block;
     margin-top: 3px;
-    color: ${props => (props.available ? '#999' : '#666')};
+    color: ${(props) => (props.available ? '#999' : '#666')};
   }
 `;
