@@ -1,25 +1,7 @@
 import axios from 'axios';
 
-import { store } from '~/store';
-
 const api = axios.create({
-  baseURL: '/api',
+  baseURL: 'http://localhost:3001/api',
 });
-
-// INTERCEPTOR FOR AUTH
-api.interceptors.request.use(
-  function(config) {
-    const { token } = store.getState().auth;
-
-    if (token !== null) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-
-    return config;
-  },
-  function(err) {
-    return Promise.reject(err);
-  }
-);
 
 export default api;
