@@ -1,7 +1,7 @@
 import storage from 'redux-persist/lib/storage';
 import { persistReducer } from 'redux-persist';
 
-export default reducers => {
+export default (reducers) => {
   const persistedReducer = persistReducer(
     {
       key: 'barberWEB',
